Uni3D/requirements.txt
 open-clip-torch
 deepspeed
 wandb 
 open3d 
 easydict 
 h5py 
 ninja
 